Conflicts:
	src/main/java/org/meteogroup/grib_library/grib2/Grib2RecordReader.java

+++ src/main/java/org/meteogroup/grib_library/util/FileChannelPartReader.java
@@ -11,7 +11,7 @@ import java.nio.channels.FileChannel;
  */
 public class FileChannelPartReader {
 
-    public byte[] readPartOfFileChannel(FileChannel fileChannel, long startPosition, int lengthToRead) throws IOException, GribReaderException {
+    public byte[] readPartOfFileChannel(FileChannel fileChannel, long startPosition, long lengthToRead) throws IOException, GribReaderException {
         if (startPosition < 0 || lengthToRead <= 0){
             throw new GribReaderException("Invalid start position or length to read");
         }
@@ -19,7 +19,7 @@ public class FileChannelPartReader {
             throw new GribReaderException("Reading to buffer will run out of FileChannel to read.");
         }
         fileChannel.position(startPosition);
-        ByteBuffer buffer = ByteBuffer.allocate(lengthToRead);
+        ByteBuffer buffer = ByteBuffer.allocate((int) lengthToRead);
         fileChannel.read(buffer);
         buffer.rewind();
         return buffer.array();
